Fix fallout from allowing impls outside of the type's definition module.
